client: Paginated based library
Overview:
Each status on the library page is independently queried at 50 entries at a
time. Entries are sorted on the 'ALL' page so that scrolling brings in entries
in the correct order.

Changes:
  - Renamed `{{library-status-button}}` to `{{library-dropdown}}`
  - Renamed route-level library components
  - Updated `ember-one-way-controls` to `v1.0.2` due to an issue they fixed.
  - Renamed pagination component to `{{pagination-scrolling}}` and moved most
    logic to a Mixin, for later creation of a manual pagination component.

Notes:
This still doesn't have full functionality due to `jsonapi-authorization` issues.
  - Cannot `CREATE` due to `record` not being fulfilled in policy.
  - Cannot `UPDATE` due to Ember-Data serializing relationships and `JA-A`
    calling `relationship#update?` which will always fail for us.
Also, the library entries are not yet filterable.

asd